refactor login/signup screens
